Разделяет js код на модули и подключает через точку входа main.js
