+++ js/dictionaries.js
@@ -1,0 +1,39 @@
+const messagePartsList = [
+  'Всё отлично!',
+  'В целом всё неплохо.',
+  'Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!'
+];
+
+const nameList = [
+  'Чернышев Тихон',
+  'Васильев Тимофей',
+  'Акимова Екатерина',
+  'Кузин Гордей',
+  'Васильев Степан',
+  'Морозова Ксения',
+  'Тихонова Валерия',
+  'Андреева Ангелина',
+  'Савельева Виктория',
+  'Лебедев Максим'
+];
+
+const descriptionList = [
+  'добрая',
+  'душевная',
+  'дружелюбная',
+  'дивная',
+  'единственная',
+  'женственная',
+  'желанная',
+  'жизнерадостная',
+  'заботливая',
+  'завораживающая'
+];
+
+export {messagePartsList, nameList, descriptionList};
